PM / freezer: Reduce freeze timeout to 1 second for Android
Freezing processes on Android usually takes less than 100 ms, and if it
takes longer than that to the point where the 20 second freeze timeout
is
reached, it's because the remaining processes to be frozen are
deadlocked
waiting for something from a process which is already frozen. There's no
point in burning power trying to freeze for that long, so reduce the
freeze
timeout to a very generous 1 second for Android and don't let anything
mess
with it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>